demo_collision - B: Bugged Hud need fix

+++ src/ecs/systems/Hud.cpp
@@ -18,8 +18,8 @@ void Hud::createHudEntities(std::vector<ECS::Entity> &gameEntities)
 {
     std::vector<std::tuple<bool, ECS::Entity &>> playerEnts;
     raylib::Window &window = raylib::Window::getWindow();
-    int scaleX = window.getWindowWidth() / 5;
-    int scaleY = window.getWindowHeight() / 7;
+    float scaleX = window.getWindowWidth() / 5;
+    float scaleY = window.getWindowHeight() / 7;
     std::string hudName;
     ECS::Entity newEnt;
     int id = 0;
@@ -35,13 +35,13 @@ void Hud::createHudEntities(std::vector<ECS::Entity> &gameEntities)
                 newEnt = Presets::createImage(hudName, {0.0, 0.0, 0.0}, {scaleX, scaleY, 0}, "../assets/hud_left.png");
                 break;
             case 2:
-                newEnt = Presets::createImage(hudName, {static_cast<float>(window.getWindowWidth() - scaleX), 0.0, 0.0}, {scaleX, scaleY, 0}, "../assets/hud_right.png");
+                newEnt = Presets::createImage(hudName, {static_cast<float>(window.getWindowWidth()) - scaleX, 0.0, 0.0}, {scaleX, scaleY, 0}, "../assets/hud_right.png");
                 break;
             case 3:
-                newEnt = Presets::createImage(hudName, {0.0, static_cast<float>(window.getWindowHeight() - scaleY), 0.0}, {scaleX, scaleY, 0}, "../assets/hud_left.png");
+                newEnt = Presets::createImage(hudName, {0.0, static_cast<float>(window.getWindowHeight()) - static_cast<float>(scaleY), 0.0}, {scaleX, scaleY, 0}, "../assets/hud_left.png");
                 break;
             case 4:
-                newEnt = Presets::createImage(hudName, {static_cast<float>(window.getWindowWidth() - scaleX), static_cast<float>(window.getWindowHeight() - scaleY), 0.0}, {scaleX, scaleY, 0}, "../assets/hud_right.png");
+                newEnt = Presets::createImage(hudName, {static_cast<float>(window.getWindowWidth()) - scaleX, static_cast<float>(window.getWindowHeight()) - static_cast<float>(scaleY), 0.0}, {scaleX, scaleY, 0}, "../assets/hud_right.png");
                 break;
             default:
                 continue;
@@ -55,8 +55,8 @@ void Hud::updateHudEntities(std::vector<ECS::Entity> &gameEntities)
 {
     std::vector<std::tuple<bool, ECS::Entity &>> hudEnts;
     raylib::Window &window = raylib::Window::getWindow();
-    int scaleX = window.getWindowWidth() / 5;
-    int scaleY = window.getWindowHeight() / 7;
+    float scaleX = window.getWindowWidth() / 5;
+    float scaleY = window.getWindowHeight() / 7;
     int id = 0;
 
     hudEnts = ECS::getEntitiesByName("hud", gameEntities);
@@ -74,7 +74,7 @@ void Hud::updateHudEntities(std::vector<ECS::Entity> &gameEntities)
                 std::get<1>(*it).getComponent<ECS::Transform>(ECS::TRANSFORM).setPosition({0.0, static_cast<float>(window.getWindowHeight()) - static_cast<float>(scaleY), 0.0});
                 break;
             case 4:
-                std::get<1>(*it).getComponent<ECS::Transform>(ECS::TRANSFORM).setPosition({static_cast<float>(window.getWindowWidth() - scaleX), static_cast<float>(window.getWindowHeight() - scaleY), 0.0});
+                std::get<1>(*it).getComponent<ECS::Transform>(ECS::TRANSFORM).setPosition({static_cast<float>(window.getWindowWidth()) - static_cast<float>(scaleX), static_cast<float>(window.getWindowHeight()) - static_cast<float>(scaleY), 0.0});
                 break;
             default:
                 continue;
